+++ lib-display/include/lcdbw.h
@@ -2,7 +2,7 @@
  * @file lcdbw.h
  *
  */
-/* Copyright (C) 2017-2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2017-2021 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -44,15 +44,15 @@ public:
 	bool Start() override;
 
 	void Cls() override;
-	void ClearLine(uint8_t) override;
+	void ClearLine(uint32_t) override;
 
 	void PutChar(int) override;
 	void PutString(const char *) override;
 
-	void Text(const char *, uint8_t);
-	void TextLine(uint8_t, const char *, uint8_t) override;
+	void Text(const char *, uint32_t);
+	void TextLine(uint32_t, const char *, uint32_t) override;
 
-	void SetCursorPos(uint8_t, uint8_t) override;
+	void SetCursorPos(uint32_t, uint32_t) override;
 	void SetCursor(uint32_t) override;
 
 private:
